+++ apps/web/src/routes/_social/-components/post-card.tsx
@@ -51,7 +51,7 @@ import {
 } from "@/lib/types";
 import {
   getActionError,
-  hasUmaminPlus,
+  hasPlusFeatures,
   isAlreadyLiked,
   isAlreadyRemoved,
   isAlreadyReposted,
@@ -313,7 +313,7 @@ export function PostCard({
       >
         <Avatar
           className={cn({
-            "avatar-shine": hasUmaminPlus(author?.createdAt),
+            "avatar-shine": hasPlusFeatures(author),
           })}
         >
           <AvatarImage src={author?.imageUrl ?? ""} alt="User avatar" />